实现子弹类及子弹更新，渲染，发射逻辑

+++ sun_bullet.h
@@ -2,6 +2,7 @@
 
 #include "bullet.h"
 #include "animation.h"
+#include "sun_bullet_ex.h"
 
 extern Atlas atlas_sun;
 extern Atlas atlas_sun_explode;
@@ -26,8 +27,8 @@ public:
 		//对齐动画帧图偏移量
 		IMAGE* frame_idle = animation_idle.get_frame();
 		IMAGE* frame_explode= animation_explode.get_frame();
-		explode_render_offset = { (frame_idle->getwidth() - frame_explode->getwidth()) / 2,
-									(frame_idle->getheight() - frame_explode->getheight()) / 2
+		explode_render_offset = { (int)((frame_idle->getwidth() - frame_explode->getwidth()) / 2),
+									(int)((frame_idle->getheight() - frame_explode->getheight()) / 2)
 		};
 
 
@@ -71,7 +72,9 @@ public:
 private:
 	Animation animation_idle;//日光子弹默认闪烁动画
 	Animation animation_explode;//日光子弹爆炸动画
-	Vector2<float> explode_render_offset;//爆炸动画渲染偏移
-	const float G=0.003f;
+	Vector2<int> explode_render_offset;//爆炸动画渲染偏移
+	const float G=0.003;
+	
+
 };
